+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # 2.32.0-0 / Unreleased
+  * [Added] `node.hasClass(className)`
   * [Added] `tree.toDict(includeRoot, callback)` and `node.toDict(recursive, callback)`:
     callback can now return `false` or `"skip"` to skip nodes.
   * [Fixed] #951 Hover issue in unselectable radio